requirements.txt
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
decorator==4.4.2
feedfinder2==0.0.4
feedparser==6.0.1
future==0.18.2
geocoder==1.38.1
geograpy3==0.1.24
idna==2.10
isodate==0.6.0
jellyfish==0.8.2
jieba3k==0.35.1
joblib==0.17.0
lxml==4.6.1
newspaper3k==0.2.8
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
Pillow==8.0.0
pycountry==20.7.3
pylodstorage==0.0.18
pyparsing==2.4.7
python-dateutil==2.8.1
python-twitter==3.5
PyYAML==5.3.1
ratelim==0.1.6
rdflib==5.0.0
regex==2020.10.15
requests==2.24.0
requests-file==1.5.1
requests-oauthlib==1.3.0
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.0.1
SPARQLWrapper==1.8.5
tinysegmenter==0.3
tldextract==2.2.3
tqdm==4.50.2
urllib3==1.25.10
